+++ .env
@@ -1,3 +1,4 @@
-USERNAME=leolee
-PASSWORD=123456
-url=
+USERNAME=15022220000
+PASSWORD=abcd1234
+url=http://192.168.0.141
+device_type=5
