--- testcases/login.yaml
@@ -1,16 +1,0 @@
-config:
-    name: login
-    base_url: http://192.168.0.141
-teststeps:
--
-    name: 登录
-    request:
-        method: POST
-        url: /index.php/login/submit
-        json:
-          username: 15011110000
-          password: abcd1234
-          device_type: "5"
-    validate:
-        - eq: [body.code, 1000]
-        - eq: [status_code, 200]
