+++ testcases/login/login_test.py
@@ -1,0 +1,54 @@
+# NOTE: Generated By HttpRunner v3.1.5
+# FROM: testcases\login.yaml
+
+
+from httprunner import HttpRunner, Config, Step, RunRequest, RunTestCase
+
+
+class TestCaseLogin(HttpRunner):
+
+    config = (
+        Config("login")
+        .variables(
+            **{
+                "name": "${ENV(USERNAME)}",
+                "pas": "${ENV(PASSWORD)}",
+                "device": "${ENV(device_type)}",
+            }
+        )
+        .base_url("${ENV(url)}")
+    )
+
+    teststeps = [
+        Step(
+            RunRequest("登录接口")
+            .post("/index.php/login/submit")
+            .with_json(
+                {"username": "$name", "password": "$pas", "device_type": "$device"}
+            )
+            .extract()
+            .with_jmespath("body.data.token", "token")
+            .validate()
+            .assert_equal("body.code", 1000)
+            .assert_equal("status_code", 200)
+        ),
+        Step(
+            RunRequest("查询积分接口")
+            .post("/index.php/activity/integralLog")
+            .with_json(
+                {
+                    "device_type": "$device",
+                    "page": 1,
+                    "limit": 1,
+                    "Authorization": "$token",
+                }
+            )
+            .validate()
+            .assert_equal("body.code", 1000)
+            .assert_equal("status_code", 200)
+        ),
+    ]
+
+
+if __name__ == "__main__":
+    TestCaseLogin().test_start()
